refactor(core): put the inbound identity next to the outbound one
A ponytail pass over the previous commit, applied. No behaviour change.

`fastmcp_server/config.py` came out of #956 holding one constant and nothing
else -- a file, a module docstring and an `__all__` for a single string. The
constant is `HANGAR_SERVER_NAME`, the INBOUND `serverInfo.name`; its counterpart
`HANGAR_CLIENT_INFO` is the OUTBOUND clientInfo, and lives in `protocol.py`,
whose comment already pointed across at the other half. They are now one module
apart no longer: both halves of "who Hangar says it is" sit together, with the
#560 history that unified them.

Also dropped: the `HANGAR_SERVER_NAME` re-export and `__all__` added to
`fastmcp_server/__init__.py` in the same commit. It had no importer -- every
consumer reaches the constant directly -- so it was flexibility invented for
nobody, in a package that had just had its exports deleted for that exact
reason.

Five import sites repointed at `mcp_hangar.protocol`.

Refs #953, #956
